accesses 3rd services via HTTPS (#1)
Found several vulnerabilities that I think are causing the issue noted on this twitter thread.

https://twitter.com/tommy/status/806479901360553984

The main domain and the scripts accessed from Gravatar, Google Fonts, and Google Analytics were all naked over the wire and thus vulnerable to a Man in the MIddle Attack. 

https://en.wikipedia.org/wiki/Man-in-the-middle_attack

Removed vulnerability by accessing `https://` instead of `http://` or `//`.